Get simple workflow working through initial find_sample_size.

+++ src/test/kotlin/org/cyrptobiotic/shangrla/core/CvrBuilder.kt
@@ -6,7 +6,7 @@ import kotlin.random.Random
 class CvrBuilders {
     val builders = mutableListOf<CvrBuilder>()
 
-    fun add(id: String, tally_pool: String? = null, pool: Boolean = false, sampled: Boolean = false, sample_num: Double? = null, p: Double? = null): CvrBuilders {
+    fun add(id: String, tally_pool: String? = null, pool: Boolean = false, sampled: Boolean = false, sample_num: Int? = null, p: Double? = null): CvrBuilders {
         val cb = CvrBuilder(id)
         cb.tally_pool = tally_pool
         cb.pool = pool
@@ -112,7 +112,7 @@ class CvrBuilder(
     var pool: Boolean = false
     var p: Double? = null
     var sampled: Boolean? = null
-    var sample_num: Double? = null
+    var sample_num: Int? = null
 
     fun setTallyPool(pool: String): CvrBuilder {
         this.tally_pool = pool
